Add sidebar/profile hook/deck list (and mock auth)
I didn't bother doing the mock useAuth very rigorously because I'm just
trying to mock it up right now using a simple react-query hook. But as
I do it, I'm not certain why this shouldn't be a valid option. Because
login and logout are both asynchronous actions with state and effects --
meaning they're mutations. And the useAuth hook doesn't contain any keys
so it should be finee to put in localStorage, and then we don't need an
auth context, just the query context.

But this is a big TODO to come back to. I think I will be happier having
the context to handle this, before too long. But for now I'm going to
move on.