Replace DaisyUI collapse and much of the main card display
This commit builds out mostly this component that is an accordion of all
the different phrases. So we're mostly replacing Daisy here, but along
the way we basically deleted the old code (as you can see) and added it
back using v0 / shadcn inspired code, so there are a few things not done
yet that are regression-y, like the "relations" section and the mutations
("add to deck", "add a new translation", "edit status", "add relation").