Add zod and tan-form to the login page
There is some TODO in here where I was not able to get the form's submit
button to properly mirror the "can we submit rn" state. Using canSubmit
doesn't work right for some reason. We're trying to use mutation.isError &&
form.state.isTouched and so on, but the form seems to _immediately_ register
as both touched and dirty and this makes it not work quite right. But it's
working; it's just the user feedback isn't as good as I would like it to be.