Move email and password to their own re-used components
This highlights one of the really nice things about react-hook-form:
When we want to call one of these input components, we are only passing
the register function, and the error state. Everything else is handled
by the library, so there's no passing of defaultValues even (handle that
inside the useForm call), and definitely no setters or handlers.

This means that the Login form and Signup form can have different rules
for how we validate and use the fields, and what we submit and how, but
the input component itself is the same; I just write a different mutation
and a different zod schema, and RHF handles the rest. Very nice.