Add the profile update form, and all its inputs
This commit completes the transition for the third and final form that
we had previously moved to tanstack forms, so it satisfies the goals of
the PR. But significantly, it is also displaying for us mostly how we
want to be handling forms and fields.

1. The term "Field" is now being used to describe an input combined
with its label, error state, controllers, etc.
1. The folder `components/fields` is being used to store these combo
components, with hardcoded labels and their own unique behaviors.
1. We start with 6 fields: email, password, username, avatar_url,
language_primary, languages_spoken.
1. The way these fields only receive "errors" and "control" so they are
just about the ui and all the logic is handled above, by the form.