Debounce the search, but not the url change
We're also removing the search terms from the query key, which
could have some performance issues but mostly will prevent us from
discarding the cached data we have every time the value changes.
But the query still re-runs because... the query function changes!
Committing now because it works, but I think, with this pattern of
having the query key stay the same and change the queryFn change,
we can just debounce this and skip an effect/state variable or two.