Make the search results show relations too
This includes one regression for overall user capabilities: the ability to unfriend
is now gone from the UI because we now use a handshake icon to show friendship that
is already established. I would like to make these avatar preview sections clickable
so you can see a public profile page for the user (or just a modal i guess) and then
this section can have the unfriend button. It can be hidden behind one extra click.