WIP: Try to send email invite
this commit attempts to use the supabase auth admin API to send an invite email,
but the supabase admin doesn't expose this to random authenticated users, so I
will have to do something else. Options include:

1. Use an API route in the repo, to run on Vercel/Cloudflare, not bundled with
    the native app, to just call the Postmark API directly, and then log the invite
    in the database.
2. Same as option 1 but use a supabase serverless function.
3. Use a link to open up the user's email; also can do Whatsapp and other links, or
    the native share dialog... this is, infrastructurally, the easiest one.

Some day we will have to add a back-end to this app, but if we go with option 3, we
can say: not today.