Refactor relationship hooks
This commit fixes the return types of the relationship hooks so that the
fetcher functions now process relationship data into the format given by
FriendSummaryRelative. This type contains fields like "profile" rather
than "profile_more" and "profile_less" -- it figures out which profile
is the relevant one, and just keeps that one and discards the other.
Likewise it has a bool for whether the current user is the uid_for or
the uid_by. So it converts the data into something that makes sense as an
object _relative to_ the logged in user.